web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask-cors>=5.0.0",
    "flask-migrate>=4.0.7",
    "flask-socketio>=5.4.1",
    "flask-sqlalchemy>=3.1.1",
    "flask[async]>=3.0.3",
    "logto>=0.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
]
